layout should be explicit in the json configuration and move LayoutSpec to layout_algorithm
